+++ src/save.c
@@ -1149,7 +1149,7 @@ int32u l;
     DEBUG(fprintf(logfile, " %02X", (int)xor_byte));
     xor_byte ^= ((l >> 24) & 0xFF);
     (void)putc((int)xor_byte, fileptr);
-    DEBUG(fprintf(logfile, " %02X = %ld\n", (int)xor_byte, (long)l));
+    DEBUG(fprintf(logfile, " %02X = %ld\n", (int)xor_byte, (int32)l));
 }
 
 static void wr_bytes(c, count)
@@ -1287,7 +1287,7 @@ int32u *ptr;
     l |= (int32u)(c ^ xor_byte) << 24;
     *ptr = l;
     DEBUG(
-        fprintf(logfile, "%02X %02X = %ld\n", (int)c, (int)xor_byte, (long)l));
+        fprintf(logfile, "%02X %02X = %ld\n", (int)c, (int)xor_byte, (int32)l));
 }
 
 static void rd_bytes(ch_ptr, count)
